+++ lib/screen_init.dart
@@ -1,15 +1,15 @@
 part of 'adaptive_screen_flutter.dart';
 
 class ScreenInit {
-  static double tabBreakPoint = 600.0;
+  static double tabBreakPoint = 600;
 
-  static double webBreakPoint = 950.0;
+  static double webBreakPoint = 950;
 
   static void setTabBreakPoint(double value) {
     tabBreakPoint = value;
   }
 
   static void setWebBreakPoint(double value) {
-    tabBreakPoint = value;
+    webBreakPoint = value;
   }
 }