+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## [0.0.1]
 
 * Initial release for adaptive screen flutter.
+
+## [0.0.2] - 02/13/2022
+
+* Bugfix web widget
+* Add example to package